extract SBRef workflows to new module
