+++ fmriprep/workflows/__init__.py
@@ -4,3 +4,4 @@
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 from .base import fmri_preprocess_single
 from .anatomical import t1w_preprocessing
+from .sbref import sbref_workflow, sbref_t1_registration
